# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /beginner/assessment:
    post:
      operationId: submitBeginnerAssessment
      tags: [practice]
      summary: Place a learner at Level 0 or Level 1
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BeginnerAssessmentInput" }
      responses:
        "200":
          description: Persisted placement
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BeginnerProfile" }
  /beginner/practice:
    post:
      operationId: recordBeginnerPractice
      tags: [practice]
      summary: Record a voice-first beginner practice attempt
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BeginnerPracticeInput" }
      responses:
        "200":
          description: Updated beginner overview
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BeginnerOverview" }
components:
  schemas:
...
    BeginnerProfile:
      type: object
      required: [level, assessmentCompleted, sessionsCompleted]
      properties:
        level: { type: string, enum: [level_0, level_1] }
        assessmentCompleted: { type: boolean }
        sessionsCompleted: { type: number }
    BeginnerAssessmentInput:
      type: object
      required: [wordRecognition, listening, repeating, pronunciation, sentenceUnderstanding, alphabetRecognition]
      properties:
        wordRecognition: { type: number, minimum: 0, maximum: 1 }
        listening: { type: number, minimum: 0, maximum: 1 }
        repeating: { type: number, minimum: 0, maximum: 1 }
        pronunciation: { type: number, minimum: 0, maximum: 1 }
        sentenceUnderstanding: { type: number, minimum: 0, maximum: 1 }
        alphabetRecognition: { type: number, minimum: 0, maximum: 1 }
        language: { type: string, enum: [hindi, roman_hindi, urdu] }
    BeginnerPracticeInput:
      type: object
      required: [itemId, kind, correct]
      properties:
        itemId: { type: string, minLength: 1 }
        kind: { type: string, enum: [word, sentence, listening, pronunciation, conversation] }
        correct: { type: boolean }
        pronunciationScore: { type: number, minimum: 0, maximum: 100 }
    BeginnerOverview:
      type: object
      required: [profile, lesson, revision, summary]
      properties:
        profile: { $ref: "#/components/schemas/BeginnerProfile" }
        lesson: { $ref: "#/components/schemas/BeginnerLesson" }
        revision:
          type: array
          items: { $ref: "#/components/schemas/BeginnerItem" }
        summary: { $ref: "#/components/schemas/BeginnerSummary" }
    BeginnerItem:
      type: object
      required: [id, kind, level, prompt, english, picture, meaning, explanation, category]
      properties:
        id: { type: string }
        kind: { type: string, enum: [word, sentence] }
        level: { type: string, enum: [level_0, level_1] }
        prompt: { type: string }
        english: { type: string }
        picture: { $ref: "#/components/schemas/BeginnerPicture" }
        meaning: { type: string }
        explanation: { type: string }
        category: { type: string }
    BeginnerPicture:
      type: object
      required: [id, icon, alt]
      properties:
        id: { type: string }
        icon: { type: string }
        alt: { type: string }
    BeginnerLesson:
      type: object
      required: [title, durationMinutes, flow, items, startItemId]
      properties:
        title: { type: string }
        durationMinutes: { type: number }
        flow:
          type: array
          items: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/BeginnerItem" }
        startItemId: { type: string }
    BeginnerSummary:
      type: object
      required: [wordsLearned, sentencesPracticed, conversationsPracticed, pronunciationMistakes, sentenceMistakes, weakItems]
      properties:
        wordsLearned: { type: number }
        sentencesPracticed: { type: number }
        conversationsPracticed: { type: number }
        pronunciationMistakes: { type: number }
        sentenceMistakes: { type: number }
        weakItems:
          type: array
          items: { type: string }
...